requirements.txt
asgiref==3.3.1
Django==3.1.7
django-cors-headers==3.7.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.6.0
Pillow==8.1.2
PyJWT==2.0.1
pytz==2021.1
sqlparse==0.4.1